Used JPA instead of pure SQL for getting cities

+++ src/main/java/ru/apashkevich/cityapp/dto/CityResponseDto.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Builder
 public class CityResponseDto {
     List<City> cities;
-    Integer total;
+    long total;
 }
